Decouple OneApplicationResource.scala
