+++ oap-cache/oap/src/main/scala/org/apache/spark/sql/hive/thriftserver/OapEnv.scala
@@ -17,12 +17,9 @@
 
 package org.apache.spark.sql.hive.thriftserver
 
-import java.io.PrintStream
-
-import org.apache.spark.{SparkConf, SparkContext}
+import org.apache.spark.SparkContext
 import org.apache.spark.internal.Logging
-import org.apache.spark.sql.{SparkSession, SQLContext}
-import org.apache.spark.sql.hive.{HiveExternalCatalog, HiveUtils}
+import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.oap.listener.OapListener
 import org.apache.spark.sql.oap.ui.OapTab
 import org.apache.spark.util.Utils
@@ -45,7 +42,6 @@ private[spark] object OapEnv extends Logging {
       sc.addSparkListener(new OapListener)
       this.sparkSession = SparkSession.getActiveSession.get
       sc.ui.foreach(new OapTab(_))
-      initialized = true
     }
   }
 }
